resolved logic to handle order for anonymous users
