+++ src/utils/utils.js
@@ -562,130 +562,3 @@ export const fetchOrderStatus = async (args, context) => {
   );
   return orders;
 };
-
-/**
- * Verify if user account is still active (Mostly for FFU users)
- *
- * Why this check?
- * FFU libraries will not sync automatically with the CULR service,
- * unlike the public libraries. This will result in old user accounts
- * not being automatically removed from CULR.
-
- * Verify that user account is NOT blocked
- *
- * Why this check?
- * Now, a 'pickupbranch' is not associated with the agency where the
- * logged-in user is currently signed in. Therefore, it's necessary
- * to verify whether the user is blocked before placing an order.
- * 
- * @param {string} props.userId context
- * @param {string} props.agencyId context
- * @param {string} props.context context
- * @param {obj} context context
- * @returns {obj} containing status, verified and statusCode
- * 
- * statusCodes:
- * 
- * OK
- * BORCHK_USER_BLOCKED_BY_AGENCY
- * BORCHK_USER_NOT_FOUND_ON_AGENCY
- *  
- */
-export async function getUserOrderAllowedStatus({ agencyId, userId }, context) {
-  // agency must be provided
-  if (!agencyId) {
-    return { status: false, statusCode: "UNKNOWN_PICKUPAGENCY" };
-  }
-
-  // Verify that the agencyId has borrowerCheck
-  const hasBorrowerCheck = await resolveBorrowerCheck(agencyId, context);
-
-  if (hasBorrowerCheck) {
-    // Fetch specific account between the loggedIn user accounts
-    const userinfo = await context.datasources.getLoader("userinfo").load({
-      accessToken: context.accessToken,
-    });
-
-    // user accounts liste
-    const accounts = userinfo.attributes.agencies;
-
-    // fetch requested account from list
-    // Local (type) account is preferred, because it will always exist
-    const account = accounts.find(
-      (a) => a.agencyId === agencyId && a.userIdType === "LOCAL"
-    );
-
-    if (!account) {
-      // No user account was found - user is not a loaner at the provided pickupbranch
-      log.warn(
-        `No account was found for user within the provided agencyId: ${JSON.stringify(
-          {
-            status: false,
-            agencyId,
-            statusCode: "UNKNOWN_USER",
-          }
-        )}`
-      );
-
-      return {
-        status: false,
-        statusCode: "UNKNOWN_USER",
-      };
-    }
-
-    const { status, blocked } = await context.datasources
-      .getLoader("borchk")
-      .load({
-        userId: account.userId,
-        libraryCode: account.agencyId,
-      });
-
-    if (blocked) {
-      // User is blocked on the provided pickupAgency
-      log.warn(
-        `User is not allowed to place an order. Borchk: ${JSON.stringify({
-          status: false,
-          agencyId,
-          blocked,
-          borchkStatus: status,
-          statusCode: "BORCHK_USER_BLOCKED_BY_AGENCY",
-        })}`
-      );
-
-      return {
-        status: false,
-        statusCode: "BORCHK_USER_BLOCKED_BY_AGENCY",
-      };
-    }
-
-    if (status !== "OK") {
-      // User does not exist on the provided pickupBranch
-      log.warn(
-        `User is not allowed to place an order. Borchk: ${JSON.stringify({
-          status: false,
-          agencyId,
-          blocked,
-          borchkStatus: status,
-          statusCode: "BORCHK_USER_NOT_FOUND_ON_AGENCY",
-        })}`
-      );
-
-      return {
-        status: false,
-        statusCode: "BORCHK_USER_NOT_FOUND_ON_AGENCY",
-      };
-    }
-
-    // Return status ok - verified
-    return { status: true, verified: true, statusCode: "OK" };
-  }
-
-  log.warn(
-    `Borchk not permitted by agencyId ${JSON.stringify({
-      agencyId,
-    })}`
-  );
-
-  // Agency does not support borrowercheck - return status ok - no verification
-  return { status: true, statusCode: "OK", verified: false };
-}
